Auto-View Inner Models (#33)
Views can track changes to "inner models"

An inner model is one contained within a parent model. For example
a list inside another list. This tracking of inner models is not
done by default on the base Model class. Instead it provides APIs
for attaching and removing inner models. The builtin model types
(e.g. List, Dict, Set, Object) do allow views to track changes
to their inner models.

A limitation of the builtin models though is the inability to know
the "path" of a change. For example in a list of list of lists. If
a change occurs in one the leaf lists, how should one learn what
the location of that list is within the larger data structure? In,
the provides models types there is no way to do this. However there
are added function 'link', 'unlink', and 'notifier' which allow
you to add this logic to your own custom models.